Edit sendEmail style
