+++ client/src/App.js
@@ -4,7 +4,6 @@ import SendEmail from './pages';
 function App() {
   return (
     <div className="App">
-      <h1>Taskillo</h1>
       <SendEmail />
     </div>
   );
