+++ client/src/components/index.js
@@ -1,3 +1,4 @@
 import { Form, SubmitButton, FormField } from './Form';
+import { Title, Text } from './sendEmail';
 
-export { Form, SubmitButton, FormField };
+export { Form, SubmitButton, FormField, Title, Text };
