+++ client/src/components/sendEmail/Text.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Text({ text, className }) {
+  return <div className={className}>{text}</div>;
+}
