+++ client/src/components/sendEmail/Title.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Title({ text, className }) {
+  return <div className={className}>{text}</div>;
+}
