+++ client/src/components/sendEmail/index.js
@@ -1,0 +1,4 @@
+import Title from './Title';
+import Text from './Text';
+
+export { Title, Text };
