+++ client/src/pages/sendEmail/index.js
@@ -1,0 +1,3 @@
+import SendEmail from './sendEmail';
+
+export default SendEmail;
