+++ client/src/pages/sendEmail/style.css
@@ -1,0 +1,18 @@
+.send-email {
+    background-color: var(--sendEmail);
+    width: 30rem;
+}
+
+.title,
+.signup-text {
+    color: #fff;
+}
+
+.signup-text {
+    text-align: justify;
+    font-size: 0.75rem;
+}
+
+.invalid {
+    color: rgb(240, 70, 70);
+}
